+++ dev/app/src/main/java/com/esms/views/contacts/ContactBox.kt
@@ -7,38 +7,28 @@ import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.unit.dp
-import coil.compose.AsyncImage
-import com.esms.R
 import com.esms.models.LocalParameters
 import com.esms.models.PhoneContact
+import com.esms.views.contacts.pic.ProfilePicture
 
 @Composable
 fun ContactBox(contact: PhoneContact) {
     val params = LocalParameters.current
-    val fallbackPainter = painterResource(id = R.drawable.ic_launcher_background)
     Row(verticalAlignment = Alignment.CenterVertically,
         modifier = Modifier
             .height(50.dp)
     ) {
         // pfp
-        AsyncImage(
-            model = contact.pfp,
-            placeholder = fallbackPainter,
-            error = fallbackPainter,
-            fallback = fallbackPainter,
-            contentDescription = params.getNicknameForNumber(contact.number, contact.name),
-            modifier = Modifier
-                .clip(CircleShape)
-                .fillMaxHeight()
+        ProfilePicture(
+            name = params.getNicknameForNumber(contact.number, contact.name),
+            number = contact.number.hashCode(),
+            url = contact.pfp
         )
         
         Spacer(modifier = Modifier.padding(horizontal = 5.dp))
